moves serializer to seperate file
